refactor: reorganise session code into separate packages
