+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/session/orchestrator/SessionSnapshotType.kt
@@ -1,4 +1,4 @@
-package io.embrace.android.embracesdk.session
+package io.embrace.android.embracesdk.session.orchestrator
 
 /**
  * Defines the states in which a session can end.
